modularizes i18n resource loader
