+++ i18n/src/main/java/de/taimos/dvalin/i18n/II18nResourceHandler.java
@@ -1,0 +1,17 @@
+package de.taimos.dvalin.i18n;
+
+/**
+ * Copyright 2018 Cinovo AG<br>
+ * <br>
+ *
+ * @author psigloch
+ */
+public interface II18nResourceHandler {
+
+    /**
+     * initialize resources
+     *
+     * @param callback the callback
+     */
+    void initializeResources(II18nCallback callback);
+}
