+++ i18n/src/test/java/de/taimos/dvalin/i18n/ResourceFrameworkTest.java
@@ -1,5 +1,6 @@
 package de.taimos.dvalin.i18n;
 
+import de.taimos.dvalin.i18n.xml.I18nXMLHandler;
 import de.taimos.dvalin.test.AbstractMockitoTest;
 import de.taimos.dvalin.test.inject.InjectionUtils;
 import org.junit.Assert;
@@ -19,7 +20,9 @@ import java.util.Locale;
  */
 public class ResourceFrameworkTest extends AbstractMockitoTest {
 
-    private II18nLoader resourceAccess = new II18nLoader();
+    private I18nLoader resourceAccess = new I18nLoader();
+
+    private I18nXMLHandler xmlHandler = new I18nXMLHandler();
 
     private static Resource findResource(String path) {
         return new ClassPathResource(path);
@@ -65,10 +68,13 @@ public class ResourceFrameworkTest extends AbstractMockitoTest {
     @Before
     public void setUp() throws NoSuchMethodException, InvocationTargetException, IllegalAccessException {
         InjectionUtils.injectValue(this.resourceAccess, "DEFAULT_LOCALE_STRING", "de");
+
         Resource xml = ResourceFrameworkTest.findResource("i18n/test.xml");
-        ResourceFrameworkTest.injectValue(this.resourceAccess, "resourceFiles", new Resource[]{xml});
+        ResourceFrameworkTest.injectValue(this.xmlHandler, "resourceFiles", new Resource[]{xml});
         Resource schema = ResourceFrameworkTest.findResource("schema/i18nSchema_v1.xsd");
-        ResourceFrameworkTest.injectValue(this.resourceAccess, "resourceSchema", new Resource[]{schema});
+        ResourceFrameworkTest.injectValue(this.xmlHandler, "resourceSchema", new Resource[]{schema});
+
+        InjectionUtils.inject(this.resourceAccess, "${i18n.type:i18n-xml}",  this.xmlHandler);
 
         Method m = this.resourceAccess.getClass().getDeclaredMethod("initializeResources");
         m.setAccessible(true);
